+++ Integrators/PhotonTracer.h
@@ -15,7 +15,7 @@ public:
 	virtual void Debug(Image* I, const Vec2i& pixel);
 	virtual void PrintStatus(std::vector<std::string>& a_Buf) const
 	{
-		double nPhotons = math::floor((double)(m_uPassesDone * w * h) / 1000000.0);
+		float nPhotons = math::floor((float)(m_uPassesDone * w * h) / 1000000.0f);
 		a_Buf.push_back(format("Photons emitted : %d[Mil]", (int)nPhotons));
 		a_Buf.push_back(format("Photons per second : %f[Mil]", nPhotons / m_fAccRuntime));
 	}
